refactor(jobs): remove unused services and update hooks paths
- Deleted outdated `jobsApi`, `useJobsParamsUpdater`, and `useJobsQueries` services.
- Moved and updated hooks to `services/jobs/hooks/` directory.
- Simplified `TeamGroupCard` rendering logic for optional `applyUrl`.
- Added `JOBS_LINK` to `MobileBottomNav` constants.
- Refactored styles and imports in job-related components for consistency.
